+++ shared/src/models/utils/occupations/exercise-occupation.ts
@@ -1,0 +1,38 @@
+import type { Type } from 'class-transformer';
+import type { Constructor } from '../../../utils';
+import { IntermediateOccupation } from './intermediate-occupation';
+import { Occupation } from './occupation';
+import { NoOccupation } from './no-occupation';
+
+export const occupations = {
+    IntermediateOccupation,
+    NoOccupation,
+};
+
+export type ExerciseOccupation = InstanceType<
+    (typeof occupations)[keyof typeof occupations]
+>;
+
+type ExerciseOccupationDictionary = {
+    [Occupation in ExerciseOccupation as Occupation['type']]: Constructor<Occupation>;
+};
+
+export type ExerciseOccupationType = ExerciseOccupation['type'];
+
+export const occupationDictionary: ExerciseOccupationDictionary = {
+    intermediateOccupation: IntermediateOccupation,
+    noOccupation: NoOccupation,
+};
+
+export const occupationTypeOptions: Parameters<typeof Type> = [
+    () => Occupation,
+    {
+        keepDiscriminatorProperty: true,
+        discriminator: {
+            property: 'type',
+            subTypes: Object.entries(occupationDictionary).map(
+                ([name, value]) => ({ name, value })
+            ),
+        },
+    },
+];