+++ shared/src/models/utils/occupations/intermediate-occupation.ts
@@ -1,0 +1,22 @@
+import { IsInt, Min } from 'class-validator';
+import { IsValue } from '../../../utils/validators';
+import { getCreate } from '../get-create';
+import type { Occupation } from './occupation';
+
+export class IntermediateOccupation implements Occupation {
+    @IsValue('intermediateOccupation')
+    readonly type = 'intermediateOccupation';
+
+    @IsInt()
+    @Min(0)
+    readonly unoccupiedUntil: number;
+
+    /**
+     * @deprecated Use static `create` method instead.
+     */
+    constructor(unoccupiedAfter: number) {
+        this.unoccupiedUntil = unoccupiedAfter;
+    }
+
+    static readonly create = getCreate(this);
+}